% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_covariance.R
\name{rolling_covariance}
\alias{rolling_covariance}
\title{Rolling covariance}
\usage{
rolling_covariance(df = 1)
}
\arguments{
\item{df}{Degrees of freedom used in the empirical estimate.}
}
\value{
A function that takes a numeric vector as argument
}
\description{
A function for creating a rolling covariance object. Beware, that the
function returned is not symmetric.
}
\examples{
a <- c(1, 2, 3, 4, 5)
b <- c(4, 2, 5, 1, 3)
c <- c(6, 4, 10, 4, 6)
covariance <- rolling_variance()

# Returns the covariance of vector 'a' and vector 'b'
covariance(a, b)

# Returns the current covariance - the covariance of vector 'a'
# and vector 'b'.
covariance()

# Returns the covariance of vector c(a, a) and vector c(b, c)
covariance(a, c)
}
