% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_mean.R
\name{rolling_mean}
\alias{rolling_mean}
\title{Rolling mean}
\usage{
rolling_mean()
}
\value{
A function that takes a numeric vector as argument
}
\description{
A function for creating a rolling mean object.
}
\examples{
a <- c(1, 2, 3, 4, 5)
b <- c(4, 2, 5, 1, 3)
mu <- rolling_mean()
mu(a) # Returns the mean of vector 'a'
mu()  # Returns the current mean - the mean of vector 'a'
mu(b) # Returns the mean of vector c(a, b)
}
